+++ concurrency/CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 ~~~~~
 
 * (:issue:`316`) ``Control.Monad.Conc.Class.unsafeUnmask``.
+* ``Control.Monad.Conc.Class.interruptible``.
 
 1.10.0.0 (2020-05-10)
 ---------------------